+++ testnet/stacks-node/src/node.rs
@@ -2,7 +2,6 @@ use std::convert::TryFrom;
 use std::default::Default;
 use std::net::SocketAddr;
 use std::sync::mpsc::{sync_channel, Receiver, SyncSender};
-use std::sync::{atomic::AtomicBool, Arc};
 use std::{collections::HashSet, env};
 use std::{thread, thread::JoinHandle, time};
 
@@ -348,7 +347,7 @@ impl Node {
         let mut event_dispatcher = EventDispatcher::new();
 
         for observer in &config.events_observers {
-            event_dispatcher.register_observer(observer, Arc::new(AtomicBool::new(true)));
+            event_dispatcher.register_observer(observer);
         }
 
         event_dispatcher.process_boot_receipts(receipts);
@@ -379,7 +378,7 @@ impl Node {
         let mut event_dispatcher = EventDispatcher::new();
 
         for observer in &config.events_observers {
-            event_dispatcher.register_observer(observer, Arc::new(AtomicBool::new(true)));
+            event_dispatcher.register_observer(observer);
         }
 
         let chainstate_path = config.get_chainstate_path_str();
@@ -841,7 +840,8 @@ impl Node {
         loop {
             let mut process_blocks_at_tip = {
                 let tx = db.tx_begin_at_tip();
-                self.chain_state.process_blocks(tx, 1)
+                self.chain_state
+                    .process_blocks(tx, 1, Some(&self.event_dispatcher))
             };
             match process_blocks_at_tip {
                 Err(e) => panic!("Error while processing block - {:?}", e),
@@ -920,17 +920,13 @@ impl Node {
             &block.header.parent_block,
         );
 
-        let chain_tip = ChainTip {
-            metadata,
-            block,
-            receipts,
-        };
-
         self.event_dispatcher.process_chain_tip(
-            &chain_tip,
+            &block,
+            &metadata,
+            &receipts,
             &parent_index_hash,
             Txid([0; 32]),
-            vec![],
+            &vec![],
             None,
             parent_burn_block_hash,
             parent_burn_block_height,
@@ -939,6 +935,11 @@ impl Node {
             &processed_block.parent_microblocks_cost,
         );
 
+        let chain_tip = ChainTip {
+            metadata,
+            block,
+            receipts,
+        };
         self.chain_tip = Some(chain_tip.clone());
 
         // Unset the `bootstraping_chain` flag.